Explicitly have instruction bundles
This allows code that depends on the BFD to decouple itself from the
fact that DREAMER currently has an odd encoding scheme that's kind-of,
sort-of like a VLIW machine, but also not quite.  The idea is to
instead just treat any Hurricane code explicitly as VLIW from tha
start, which should make the simulator easier to paramaterize for this
behavior.