Rework operand conversion
This makes the "idx()" and "dir()" functions look like the "lit()"
function.  There's two changes here:

* These functions are virtual (and abstract) members of operand

* These functions return a tagged union with False/Unit

The general idea here is that rather than having to do something like
a RTTI cast to figure out what sort of operand something is, you just
try to parse it as a literal (or whatever).  This was the way it was
originally designed.

Apparently I forgot about that though, which is why this also removes
the net() function...