Add special parallel network bundling logic
The old bundling logic just OR'd all instructions together, but that
doesn't work at all with parallel network operations.  It turns out
there's only two of those in the Histogram benchmark, but this ends up
being necessary to make that work.

Note that it's still a bit flaky:

* This only works with 2-instruction bundles (so no network fanout)
* The network op must be the second operation

This can all be fixed.